three-tier-app/backend/requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
PyMySQL==1.0.3
SQLAlchemy==2.0.24
python-dotenv==1.0.0
bcrypt==4.0.1
PyJWT==2.8.0
gunicorn==20.1.0
